Added variable queuing ability.

+++ src/Hardware.h
@@ -34,8 +34,8 @@ public:
   void     turnOffHeaters();
   void     setHeaterMode(bool on);
 
-  void     queueValve(uint32_t  duration, bool real);
-  void     queueBallast(uint32_t  duration, bool real);
+  void     queueValve(uint16_t  duration, bool real);
+  void     queueBallast(uint16_t  duration, bool real);
   void     clearValveQueue();
   void     clearBallastQueue();
   bool     checkValve();
